+++ samples/async_gasg.py
@@ -129,15 +129,27 @@ def sdk_version():
     print(f"async_gas_version: {_async_gas_version} SDK version: {__version__}")
 
 
+def _group_pull(cli_arg: list, config: PysuiConfiguration) -> list:
+    """Check for group and set accordingly."""
+
+    if cli_arg.count("--group"):
+        ndx = cli_arg.index("--group")
+        nvl: str = cli_arg[ndx + 1]
+        config.make_active(group_name=nvl)
+
+    return cli_arg
+
+
 def main():
     """Setup asynch loop and run."""
-    arg_line = sys.argv[1:].copy()
     # Handle a different client.yaml than default
-    if arg_line and arg_line[0] == "--local":
-        raise ValueError("--local is invalid for GraphQL")
     cfg = PysuiConfiguration(group_name=PysuiConfiguration.SUI_GQL_RPC_GROUP)
+    arg_line = _group_pull(sys.argv[1:].copy(), cfg)
     parsed = build_async_gas_parser(arg_line, cfg)
-    cfg.make_active(profile_name=parsed.profile_name, persist=False)
+
+    cfg.make_active(
+        group_name=parsed.group_name, profile_name=parsed.profile_name, persist=False
+    )
     if parsed.version:
         sdk_version()
         return