+++ samples/cmd_argsg.py
@@ -735,24 +735,20 @@ def _base_parser(
     pconfig: PysuiConfiguration, desc: Optional[str] = None
 ) -> argparse.ArgumentParser:
     """Basic parser setting for all commands."""
-    # if desc:
-    #     desc = "%(prog)s " + desc
-    # else:
-    #     desc = "%(prog)s [options] command [--command_options]"
     parser = argparse.ArgumentParser(
         add_help=True,
         usage=desc or "%(prog)s [options] command [--command_options]",
         description="",
         formatter_class=argparse.RawDescriptionHelpFormatter,
     )
-    # parser.add_argument(
-    #     "--group",
-    #     dest="group_name",
-    #     choices=pconfig.group_names(),
-    #     default=pconfig.active_group,
-    #     required=False,
-    #     help=f"The GraphQL groups. Default to '{pconfig.active_group_name}'",
-    # )
+    parser.add_argument(
+        "--group",
+        dest="group_name",
+        choices=pconfig.group_names(),
+        default=pconfig.active_group.group_name,
+        required=False,
+        help=f"The GraphQL groups. Default to '{pconfig.active_group_name}'",
+    )
     parser.add_argument(
         # "-p",
         "--profile",
@@ -760,7 +756,7 @@ def _base_parser(
         choices=pconfig.profile_names(),
         default=pconfig.active_profile,
         required=False,
-        help=f"The GraphQL profile representing target GraphQL node. Default to '{pconfig.active_profile}'",
+        help=f"The GraphQL profile node. Default to '{pconfig.active_profile}' from '{pconfig.active_group_name}'",
     )
     parser.add_argument(
         "-v",