+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `object_changes()` to `TransactionResultGQL` returns data structure of changes.
+
 ### Fixed
 
 - [bug](https://github.com/FrankC01/pysui/issues/266) Fixes exception, return error if budget greater than balance